requirements.txt
astroid==2.4.2
attrs==20.3.0
cycler==0.10.0
Django==2.2
djangorestframework==3.12.2
importlib-metadata==3.1.0
iniconfig==1.1.1
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.4.0
